+++ CHANGES.md
@@ -5,6 +5,8 @@
 * remove `.pbf` extension in tiles endpoints
 * add `orjson` as an optional dependency (for faster JSON encoding/decoding within the database communication)
 * enable `geom` query parameter to select the `geometry column` (defaults to the first one)
+* add FastAPI application `exception handler` in default app
+* add `CacheControlMiddleware` middleware
 
 **breaking changes**
 
@@ -13,6 +15,7 @@
 * `table_catalog` is now of `Dict[str, Dict[str, Any]]` type (instead of `List[Dict[str, Any]]`)
 * renamed `timvt.db.table_index` to `timvt.dbmodel.get_table_index`
 * default to only view tables within the `public` schema
+* renamed *base exception class* to `TiMVTError`
 
 ## 0.7.0 (2022-06-09)